requirements.txt
pytest
pandas
google-genai
